+++ mike_ehrmantraut_bot/main.py
@@ -1,16 +1,14 @@
-import discord
-from discord import app_commands
-from discord.ext import commands
-from dotenv import load_dotenv
-from webserver import keep_alive
 import os
-from commands import setup_commands
 
-load_dotenv()
+from discord.ext import commands
+import discord
+
+from mike_ehrmantraut_bot.commands import setup_commands
 
 BOT_TOKEN = os.getenv("TOKEN")
 
 bot = commands.Bot(command_prefix="!", intents=discord.Intents.all())
+
 
 @bot.event
 async def on_ready():
@@ -21,11 +19,11 @@ async def on_ready():
         print(f"Synced {len(synced)} command(s)")
     except Exception as e:
         print(e)
-    
+
     num_servers = len(bot.guilds)
     print(f"This bot is in {num_servers} servers!")
-  
+
+
 setup_commands(bot)
 
-keep_alive()
 bot.run(BOT_TOKEN)
