+++ pyproject.toml
@@ -1,32 +1,16 @@
 [tool.poetry]
-name = "python-template"
+name = "mike-ehrmantraut-bot"
 version = "0.1.0"
 description = ""
-authors = ["Your Name <you@example.com>"]
+authors = ["dacoder101 <bobbye23899@gmail.com>"]
+readme = "README.md"
 
 [tool.poetry.dependencies]
-python = ">=3.10.0,<3.11"
-Flask = "^2.2.0"
-discord = "^2.3.2"
-python-dotenv = "^1.0.0"
-webserver = "^0.10.3"
-requests = "^2.31.0"
+python = "^3.11"
 discord-py = "^2.3.2"
+requests = "^2.31.0"
 
-[tool.poetry.dev-dependencies]
-debugpy = "^1.6.2"
-urllib3 = "1.26.15"
-poetry = {url = "https://storage.googleapis.com/poetry-bundles/poetry-1.1.15-py2.py3-none-any.whl"}
-
-[tool.pyright]
-# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
-useLibraryCodeForTypes = true
-
-[tool.ruff]
-# https://beta.ruff.rs/docs/configuration/
-select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
-ignore = ['W291', 'W292', 'W293']
 
 [build-system]
-requires = ["poetry-core>=1.0.0"]
+requires = ["poetry-core"]
 build-backend = "poetry.core.masonry.api"
