--- webserver.py
@@ -1,26 +1,0 @@
-from flask import Flask
-
-from threading import Thread
-
-app = Flask('')
-
-
-@app.route('/')
-def home():
-
-  return "I'm alive (just a placeholder)"
-
-
-def run():
-
-  app.run(host='0.0.0.0', port=8080)
-
-
-def keep_alive():
-
-  t = Thread(target=run)
-
-  t.start()
-
-
-#discord profile connection
